LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-keyboard-navigation-order-expected.txt










FAIL Radio button keyboard navigation should proceed in tree-order. assert_equals: Navigated to next radio button in form 'mix-form' expected Element node <input id="mix-formNew" type="radio" form="mix-form" name... but got Element node <input type="radio" name="mix-form" id="mix-form2"></input>
